pyproject.toml
[project]
name = "mkslides-mcp-server"
version = "0.1.0"
description = "MCP server for generating slides using mkslides"
requires-python = ">=3.12"
dependencies = [
    "mcp",
    "PyYAML",
    "mkslides"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Optional: Define scripts if you want a named entry point like mcp-server-git
# [project.scripts]
# mkslides-server = "mkslides_server:main" # Assumes a main function in your script

[tool.setuptools.packages.find]
where = ["src"] # list of folders that contain the packages (["."])
include = ["mkslides_mcp_server*"] # package names should match these glob patterns (["*"])
# exclude = [] # exclude packages matching these glob patterns ([] )
# namespaces = false # to disable scanning PEP 420 namespaces (true)
